refactored WindowInputNode to hosting, changed nodes accordingly, exposed some more features
