+++ vvvv45/src/core/Hosting/Hosting.csproj
@@ -125,6 +125,7 @@
     <Compile Include="VVVVLogger.cs" />
     <Compile Include="Factories\FreeFrameFactory.cs" />
     <Compile Include="Factories\VSTFactory.cs" />
+    <Compile Include="WindowInputNode.cs" />
   </ItemGroup>
   <ItemGroup />
   <ItemGroup>
