+++ vvvv45/src/nodes/plugins/NativePlugins.csproj
@@ -92,7 +92,6 @@
     <Compile Include="System\LegacyKeyboardHelper.cs" />
     <Compile Include="System\MouseNode.cs" />
     <Compile Include="System\SystemProcessorsNode.cs" />
-    <Compile Include="System\WindowInputNode.cs" />
     <Compile Include="Texture\ImagePlayer\FrameDecoder.cs" />
     <Compile Include="Texture\ImagePlayer\FrameInfo.cs" />
     <Compile Include="Texture\ImagePlayer\ImagePlayer.cs" />
